2014-12-03 Roman - FIX:SYS: Some build warnings fixed - FIX:UI.{VCAEngine,Vision}: Changes down procedure mostly fixed and adapted for update the parent widget into open for edit state.
git-svn-id: svn://oscada.org/trunk/OpenSCADA@2177 7bc60720-51fb-0310-bb35-cf765b5d350c

+++ src/tsys.cpp
@@ -1033,7 +1033,7 @@ string TSYS::strEncode( const string &in, TSYS::Code tp, const string &opt1 )
 	case TSYS::Bin: {
 	    string svl, evl;
 	    sout.reserve(in.size()/2);
-	    for(unsigned iCh = 0; iCh < (int)in.size(); ++iCh)
+	    for(unsigned iCh = 0; iCh < in.size(); ++iCh)
 		if(isxdigit(in[iCh])) {
 		    sout += (char)strtol(in.substr(iCh,2).c_str(),NULL,16); iCh++;
 		}
